+++ R/adaptive.bin.R
@@ -46,7 +46,7 @@ increment_counter <- function(pointers, that.n){
 #' 
 #' This is an internal function. It creates EICs using adaptive binning procedure
 #' 
-#' @param x A matrix with columns of m/z, retention time, intensity.
+#' @param features A tibble with columns of m/z, retention time, intensity.
 #' @param min.run Run filter parameter. The minimum length of elution time for a series of signals grouped by m/z to be 
 #'  considered a peak.
 #' @param min.pres Run filter parameter. The minimum proportion of presence in the time period for a series of signals grouped 
@@ -62,9 +62,9 @@ increment_counter <- function(pointers, that.n){
 #' \itemize{
 #'   \item height.rec - The records of the height of each EIC.
 #'   \item masses - The vector of m/z values after binning.
-#'   \item labels - The vector of retention time after binning.
+#'   \item rt - The vector of retention time after binning.
 #'   \item intensi - The vector of intensity values after binning.
-#'   \item grps - The EIC labels, i.e. which EIC each observed data point belongs to.
+#'   \item grps - The EIC rt, i.e. which EIC each observed data point belongs to.
 #'   \item times - All the unique retention time values, ordered.
 #'   \item tol - The m/z tolerance level.
 #'   \item min.count.run - The minimum number of elution time points for a series of signals grouped by m/z to be considered a peak.
@@ -72,37 +72,36 @@ increment_counter <- function(pointers, that.n){
 #' @export
 #' @examples
 #' adaptive.bin(raw.data, min.run = min.run, min.pres = min.pres, tol = tol, baseline.correct = baseline.correct, weighted = intensity.weighted)
-adaptive.bin <- function(x,
-                         min.run,
-                         min.pres,
-                         tol,
-                         baseline.correct,
-                         weighted = FALSE) {
+adaptive.bin <- function(features,
+                         min_elution_length,
+                         min_presence,
+                         mz_tol,
+                         baseline_correct,
+                         intensity_weighted = FALSE) {
   # order inputs after mz values
-  data_table <- tibble::tibble(mz = x$masses, labels = x$labels, intensities = x$intensi) |> dplyr::arrange_at("mz")
+  features <- features |> dplyr::arrange_at("mz")
 
 
-  cat(c("m/z tolerance is: ", tol, "\n"))
+  cat(c("m/z tolerance is: ", mz_tol, "\n"))
 
-  times <- x$times #sort(unique(data_table$labels))
+  times <- sort(unique(features$rt))
 
-  rm(x)
   min_time <- min(times)
   max_time <- max(times)
   time_range <- (max_time - min_time)
 
   # calculate function parameters
-  min.count.run <- min.run * length(times) / time_range
+  min.count.run <- min_elution_length * length(times) / time_range
   aver.time.range <- (time_range) / length(times)
 
   # init data
-  newprof <- matrix(0, nrow = length(data_table$mz), ncol = 4)
-  height.rec <- matrix(0, nrow = length(data_table$mz), ncol = 3)
+  newprof <- matrix(0, nrow = length(features$mz), ncol = 4)
+  height.rec <- matrix(0, nrow = length(features$mz), ncol = 3)
 
   # init counters
   pointers <- list(curr.label = 1, prof.pointer = 1, height.pointer = 1)
 
-  breaks <- compute_breaks(tol, data_table$mz, data_table$intensities, weighted)
+  breaks <- compute_breaks(mz_tol, features$mz, features$intensities, intensity_weighted)
 
   for (i in 1:(length(breaks) - 1))
   {
@@ -111,12 +110,12 @@ adaptive.bin <- function(x,
     start <- breaks[i] + 1
     end <- breaks[i + 1]
 
-    this_table <- data.frame(labels = data_table$labels[start:end], mz = data_table$mz[start:end], intensities = data_table$intensities[start:end])
+    this_table <- dplyr::slice(features, (start:end))
 
-    if (length(unique(this_table$labels)) >= min.count.run * min.pres) {
-      # reorder in order of labels (scan number)
-      this_table <- this_table |> dplyr::arrange_at("labels")
-      mass.den <- compute_densities(this_table$mz, tol, weighted, this_table$intensities, median)
+    if (length(unique(this_table$rt)) >= min.count.run * min_presence) {
+      # reorder in order of rt (scan number)
+      this_table <- this_table |> dplyr::arrange_at("rt")
+      mass.den <- compute_densities(this_table$mz, mz_tol, intensity_weighted, this_table$intensities, median)
 
       mass.den$y[mass.den$y < min(this_table$intensities) / 10] <- 0
       mass.turns <- find.turn.point(mass.den$y)
@@ -138,17 +137,17 @@ adaptive.bin <- function(x,
 
         if (nrow(that) > 0) {
           that <- combine.seq.3(that) |> dplyr::arrange_at("mz")
-          that.range <- diff(range(that$labels))
+          that.range <- span(that$rt)
 
-          if (that.range > 0.5 * time_range & length(that$labels) > that.range * min.pres & length(that$labels) / (that.range / aver.time.range) > min.pres) {
-            that$intensities <- rm.ridge(that$labels, that$intensities, bw = max(10 * min.run, that.range / 2))
+          if (that.range > 0.5 * time_range & length(that$rt) > that.range * min_presence & length(that$rt) / (that.range / aver.time.range) > min_presence) {
+            that$intensities <- rm.ridge(that$rt, that$intensities, bw = max(10 *min_elution_length, that.range / 2))
 
             that <- that |> dplyr::filter(intensities != 0)
           }
 
           that.n <- length(that$mz)
 
-          newprof[pointers$prof.pointer:(pointers$prof.pointer + that.n - 1), ] <- cbind(that$mz, that$labels, that$intensities, rep(pointers$curr.label, that.n))
+          newprof[pointers$prof.pointer:(pointers$prof.pointer + that.n - 1), ] <- cbind(that$mz, that$rt, that$intensities, rep(pointers$curr.label, that.n))
           height.rec[pointers$height.pointer, ] <- c(pointers$curr.label, that.n, max(that$intensities))
 
           # increment counters
@@ -157,12 +156,12 @@ adaptive.bin <- function(x,
       }
     } else {
       if (runif(1) < 0.05) {
-        this_table <- this_table |> dplyr::arrange_at("labels")
+        this_table <- this_table |> dplyr::arrange_at("rt")
 
         that.merged <- combine.seq.3(this_table)
         that.n <- nrow(that.merged)
 
-        newprof[pointers$prof.pointer:(pointers$prof.pointer + that.n - 1), ] <- cbind(that.merged$mz, that.merged$labels, that.merged$intensities, rep(pointers$curr.label, that.n))
+        newprof[pointers$prof.pointer:(pointers$prof.pointer + that.n - 1), ] <- cbind(that.merged$mz, that.merged$rt, that.merged$intensities, rep(pointers$curr.label, that.n))
         height.rec[pointers$height.pointer, ] <- c(pointers$curr.label, that.n, max(that.merged$intensities))
 
         # increment counters
@@ -176,14 +175,11 @@ adaptive.bin <- function(x,
 
   newprof <- newprof[order(newprof[, 1], newprof[, 2]), ]
 
+  newprof_tibble <- tibble::tibble(mz = newprof[, 1], rt = newprof[, 2], intensities = newprof[, 3], grps = newprof[, 4])
+
   raw.prof <- new("list")
   raw.prof$height.rec <- height.rec
-  raw.prof$masses <- newprof[, 1]
-  raw.prof$labels <- newprof[, 2]
-  raw.prof$intensi <- newprof[, 3]
-  raw.prof$grps <- newprof[, 4]
-  raw.prof$times <- times
-  raw.prof$tol <- tol
+  raw.prof$features <- newprof_tibble
   raw.prof$min.count.run <- min.count.run
 
   return(raw.prof)